requirements.txt
fastapi
uvicorn
pydantic
requests
edgartools<3
langchain
langgraph
langchain-openai
langchain-google-genai
python-dotenv
celery[redis]
redis
psycopg[binary]
psycopg_pool
